// src/lib/database.types.ts
export interface Database {
  public: {
    Tables: {
      products: {
        Row: {
          id: string
          created_at: string
          name: string
          description: string
          price: number
          category_id: string
          image_urls: string[]
          image_url: string
          sizes: string[]
          colors: string[]
          stock: number
          is_featured: boolean
          discount_price?: number
        }
        Insert: {
          id?: string
          created_at?: string
          name: string
          description: string
          price: number
          category_id: string
          image_urls: string[]
          sizes: string[]
          colors: string[]
          stock: number
          is_featured?: boolean
          discount_price?: number
        }
        Update: {
          id?: string
          created_at?: string
          name?: string
          description?: string
          price?: number
          category_id?: string
          image_urls?: string[]
          sizes?: string[]
          colors?: string[]
          stock?: number
          is_featured?: boolean
          discount_price?: number
        }
      }
      categories: {
        Row: {
          id: string
          created_at: string
          name: string
          slug: string
          image_url: string
        }
        Insert: {
          id?: string
          created_at?: string
          name: string
          slug: string
          image_url?: string
        }
        Update: {
          id?: string
          created_at?: string
          name?: string
          slug?: string
          image_url?: string
        }
      }
// ...
      reviews: {
        Row: {
          id: string
          user_id: string
          product_id: string
          rating: number
          comment: string
          created_at: string
        }
        Insert: {
          id?: string
          user_id: string
          product_id: string
          rating: number
          comment: string
          created_at?: string
        }
        Update: {
          id?: string
          user_id?: string
          product_id?: string
          rating?: number
          comment?: string
          created_at?: string
        }
      },
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
  }
}